:dizzy: /api/join → routes/api/join.js
